moves test from servoshell file
